requirements.txt
backports.functools-lru-cache==1.6.1
cycler==0.10.0
kiwisolver==1.1.0
lxml==4.5.0
matplotlib==2.2.5
numpy==1.16.6
opencv-python==4.2.0.32
openslide-python==1.1.1
pandas==0.24.2
Pillow==6.2.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
